requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pandas==2.0.3
SQLAlchemy==2.0.19